requirements.txt
streamlit==1.37.1
langchain==0.2.14
langgraph==0.2.28
faiss-cpu==1.8.0
rank-bm25==0.2.2
pypdf==4.3.1
python-dotenv==1.0.1
pandas==2.2.2
tqdm==4.66.4
openai>=1.50.0
groq>=0.9.0
numpy
requests
